backend/pyproject.toml
[project]
name = "swasthya-api"
version = "1.0.0"
description = "Swasthya AI-Powered Medical Platform API"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-jwt-extended>=4.6.0",
    "flask-cors>=4.0.0",
    "flask-migrate>=4.0.5",
    "pymysql>=1.1.0",
    "cryptography>=41.0.0",
    "g4f>=0.2.0",
    "python-dotenv>=1.0.0",
    "werkzeug>=3.0.0",
    "gunicorn>=21.2.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.1.0",
]
